rename graphql constants file
