+++ strategies/graphQL.py
